regulator: s2dos05: add Samsung s2dos05 driver
Add driver for s2dos05 regulator.
The s2dos05 is a companion power management IC for the smart phones,
and provides LDOs[1~4] and BUCKs[1].

Signed-off-by: Dzmitry Sankouski <dsankouski@gmail.com>
Cc: Liam Girdwood <lgirdwood@gmail.com>
Cc: Mark Brown <broonie@kernel.org>

Series-changes: 2
- remove interrupt specific code, because it's
empty in vendor kernel, and I cannot test it on
available hardware anyway.

Series-version: 2
Series-to: linux-kernel@vger.kernel.org
Cover-letter:
This is continued work on Samsung S9(SM-9600) starqltechn

- Samsung s2dos05 pmic driver added
- touchscreen support
- usb 2.0 support
- some cleanup

END