arm64: dts: qcom: sdm845-db845c: mark pcie0_3p3v_dual always on
This regulator is also required to enable the USB type-A VBUS. The dwc3
driver doesn't support any regulator supplies, and as this is a dev
board we don't need to worry about power consumption. Mark this
regulator as always-on so that the left USB port (which is connected
directly to the secondary USB controller) is functional even when PCIe
is disabled.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>